Soil Temperature Modification
Modified code to use two methods of estimating soil temperature, which is used as a proxy for the temperature of baseflow. The original method uses a daily running mean of annual air temperature and the second method uses a smoothing function of daily air temperature (Tsoil(t) = (1-kappa)*Tsoil(t-1) + kappa*Tair(t)). The choice of annual air temperature ('annualT') requires a separate forcing file (assumed to be a daily running mean of annual air temperature). The choice of smoothing ('smoothT') requires specification of the smoothing parameter kappa and an initial soil temperature map (or a default spatially uniform value, which is hardcoded). If choice is 'smoothT' code also dumps soil temperature to state. Choice of soil temperature method, annual air temperature input file, soil temperature initial state, and kappa are specified in the configuration file.